+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## Version 1.37
 - data export: fix filter operator behaviour with null
 - data import : support hard delete  [issue 129](https://github.com/dufoli/Salesforce-Inspector-Advanced/129)
+- data import : support Bulk API [issue267](https://github.com/dufoli/Salesforce-Inspector-Advanced/267)
 - metadata: automatic download produce zip file and not csv and add metadataStatus json file [issue 262](https://github.com/dufoli/Salesforce-Inspector-Advanced/262)
 - add a report bug button on popup [issue 264](https://github.com/dufoli/Salesforce-Inspector-Advanced/264)